+++ src/scholarship.js
@@ -18,7 +18,8 @@ function scholarships( data ) {
   data.pell = enforceRange( data.pell, 0, data.pellMax );
 
   // Total Grants
-  data.grantsTotal = data.pell + data.scholarships;
+  data.grantsTotal = data.pell + data.scholarships +
+                     data.GIBill + data.militaryTuitionAssistance;
 
   // First Year Net Cost
   data.firstYearNetCost = data.yearOneCosts - data.grantsTotal;
@@ -29,8 +30,9 @@ function scholarships( data ) {
                       data.state529plan +
                       data.workstudy;
 
-  // grants and savings
-  data.totalgrantsandsavings = data.savingsTotal + data.grantsTotal;
+  // Total grants and savings
+  data.grantsSavingsTotal = data.grantsTotal + data.savingsTotal;
+
   return data;
 }