+++ CHANGELOG.md
@@ -1,7 +1,24 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
-## 0.1.1 2016-01-08
+## 1.0.2 2016-01-20
+- Change to remainingCost calculation
+
+## 1.0.1 2016-01-20
+- Fixed incorrect calculations of totals
+- Removed old reference to totalgrantsandsavings (now grantsSavingsTotal)
+
+## 1.0.0 2016-01-20
+- Changes to many variable names (reason for version change)
+- Added GIBill and militaryTuitionAssist as parameters, included them in grantsTotal
+- Fix for private loan total
+- added test for multiple private loan totals
+- References to Stafford loan (staff) updated to Direct loan
+
+## 0.1.2 2016-01-13
+- fixed bug with totals of multiple private loans
+
+## 0.1.1 2016-01-13
 - added deferPeriod to privateLoanMulti (each private loan can have an individual deferral period)
 
 ## 0.1.0 2016-01-08